Solved both problems mentioned in commit 8. Used an auxiliary list to make sure I didn't add duplicate entries to the Visitation Queue, and then used that list to unmark visited nodes after the algorithm has finished running.
I've added a text file and an image file to the repo's root directory so that you can see the output I got from the test graph I sketched out.

The last test case on there is still a fail case, and will be solved once I deal with adding access to an edge from *both* vertices connect to it, rather than just the 'source' node.

This commit will also keep the testing functions and outputs so that anyone trying to replicate my output can do so successfully, or just see the output on their own graphs... not that I really expect anyone to be watching this repo.