Updated it so that the edges are 'two-way' -- really, I just added another edge onto the 'destination' node that refers back to the 'source' node.
It's not necessarily best practice, but I doubt a number of the things I've done so far are. This is just a random, fun-time project anyways.

Right, so next I'll test out my current functionality on a few harder graphs. Then I'll think about the following:

1) DFS
2) Distinctions for directed graphs
3) Making edge weights impactful.
4) Naming and stating the shortest path located by BFS.
5) Trees and variants thereof (BSTs, Syntax Trees, etc.)
6) Other graph algorithms that I should incorporate.
7) Restructing the classes and project folders.

And some other stuff.