Bit of success achieved with getting the path, not fully done yet.
It looks for the first path it can find, from left to right, starting at whichever node you begin with. So basically, it works like a DFS and just terminates. I can always just rebrand this as a DFS (lol) and run the function multiple times under a while loop until an external boolean flag (passed to function by reference) is triggered to break the loop when all immediate neighbours of the starting node have been traversed. Could also stick all that within a recursive method so that neighbors with multiple children would have all of their neighbours explored as well; then that'd be a DFS done, pretty much.

I need to change it to first find all the possible paths, then return the one with the least weight. So there's still work to be done.

I'm also not entirely sure why it works, since I got tired of thinking and just started typing in code without wondering what it did (lol). I have some ideas now with respect to storing multiple possible paths when found.